Move Nebius placeholder cleanup to standardization
